requirements.txt
enum34>=1.1
Pillow>=7.1
pytz
requests>=2.23
pymongo
aiogram
aiohttp[speedups]
bcrypt==3.1.7
coverage==5.4
